% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointsV2.R
\name{plotBreakpointsV2}
\alias{plotBreakpointsV2}
\title{Plot one series with the changes found by analyseBreakpointsV2()}
\usage{
plotBreakpointsV2(dates, x, bp, y_lab = NULL)
}
\arguments{
\item{dates, x}{The series.}

\item{bp}{Rows of \code{\link[=analyseBreakpointsV2]{analyseBreakpointsV2()}} / \code{\link[=analyseClusterBreakpoints]{analyseClusterBreakpoints()}}
for this series (may be 0 rows).}

\item{y_lab}{Y-axis label.}
}
\value{
A ggplot object.
}
\description{
Plot one series with the changes found by analyseBreakpointsV2()
}
\seealso{
Other breakpoint analysis v2:
\code{\link[=analyseBreakpointsV2]{analyseBreakpointsV2()}},
\code{\link[=analyseClusterBreakpoints]{analyseClusterBreakpoints()}},
\code{\link[=breakpointContext]{breakpointContext()}},
\code{\link[=flagCommonBreaks]{flagCommonBreaks()}},
\code{\link[=inspectBreakpoint]{inspectBreakpoint()}},
\code{\link[=plotInspection]{plotInspection()}},
\code{\link[=rankBreakpoints]{rankBreakpoints()}},
\code{\link[=spatialMonthTable]{spatialMonthTable()}},
\code{\link[=vbAgeMonths]{vbAgeMonths()}}
}
\concept{breakpoint analysis v2}
